# BilfenAppPhp.RestAPI/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bilfen API",
        "version": "v1"
    },
    "components": {
        "schemas": {
            "Document": {
                "type": "object",
                "required": ["Name", "FilePath", "Status", "CreatedAt", "UpdatedAt"],
                "properties": {
                    "Id": { "type": "integer", "description": "The auto-generated id of the document" },
                    "Name": { "type": "string", "description": "The document's name" },
                    "FilePath": { "type": "string", "description": "The document's file path" },
                    "Status": { "type": "string", "description": "The document's status" },
                    "CreatedAt": { "type": "string", "format": "date-time", "description": "The document's created date" },
                    "UpdatedAt": { "type": "string", "format": "date-time", "description": "The document's updated date" },
                    "UserId": { "type": "integer", "description": "The document's user id" },
                    "User": { "$ref": "#/components/schemas/User" }
                },
                "example": {
                    "Id": 0,
                    "Name": "string",
                    "FilePath": "string",
                    "Status": "string",
                    "CreatedAt": "string",
                    "UpdatedAt": "string",
                    "UserId": 0,
                    "User": {
                        "Id": 0,
                        "FirstName": "string",
                        "LastName": "string",
                        "UserName": "string",
                        "Email": "string",
                        "Password": "string",
                        "PasswordHash": "string",
                        "createdAt": "string",
                        "updatedAt": "string",
                        "RoleId": 0,
                        "Role": {
                            "Id": 0,
                            "Name": "string",
                            "createdAt": "string",
                            "updatedAt": "string"
                        }
                    }
                }
            },
            "User": {
                "type": "object",
                "required": ["FirstName", "LastName", "UserName", "Email", "Password", "PasswordHash", "createdAt", "updatedAt",
                "RoleId", "TCKN", "MotherName", "FatherName", "BirthDate", "Gender", "CivilStatus", "EmploymentDate", "CountryId",
                "CityId", "DistrictId", "Address"],
                "properties": {
                    "Id": { "type": "integer", "description": "The auto-generated id of the user" },
                    "FirstName": { "type": "string", "description": "The user's first name" },
                    "LastName": { "type": "string", "description": "The user's last name" },
                    "UserName": { "type": "string", "description": "The user's user name" },
                    "Email": { "type": "string", "format": "email", "description": "The user's email" },
                    "Password": { "type": "string", "description": "The user's password" },
                    "PasswordHash": { "type": "string", "description": "The user's hashed password" },
                    "ImagePath": { "type": "string", "description": "The user's image path" },
                    "TCKN": { "type": "string", "description": "The user's image path" },
                    "MotherName": { "type": "string", "description": "The user's mother name" },
                    "FatherName": { "type": "string", "description": "The user's father name" },
                    "BirthDate": { "type": "string", "description": "The user's birth date" },
                    "Gender": { "type": "string", "description": "The user's gender" },
                    "CivilStatus": { "type": "string", "description": "The user's civil status" },
                    "EmploymentDate": { "type": "string", "description": "The user's employment date" },
                    "MilitaryStatus": { "type": "string", "description": "The user's military status" },
                    "PostponementDate": { "type": "string", "description": "The user's postponement date" },
                    "Address": { "type": "string", "description": "The user's postponement date" },
                    "createdAt": { "type": "string", "format": "date-time", "description": "The user's created date" },
                    "updatedAt": { "type": "string", "format": "date-time", "description": "The user's updated date" },
                    "RoleId": { "type": "integer", "description": "The user's role id" },
                    "Role": { "$ref": "#/components/schemas/Role" },
                    "CountryId": { "type": "string", "description": "The user's country id" },
                    "Country": { "$ref": "#/components/schemas/Country" },
                    "CityId": { "type": "string", "description": "The user's city id" },
                    "City": { "$ref": "#/components/schemas/City" },
                    "DistrictId": { "type": "string", "description": "The user's district id" },
                    "District": { "$ref": "#/components/schemas/District" }
                },
                "example": {
                    "Id": 0,
                    "FirstName": "string",
                    "LastName": "string",
                    "UserName": "string",
                    "Email": "string",
                    "Password": "string",
                    "PasswordHash": "string",
                    "ImagePath": "string",
                    "TCKN": "string",
                    "MotherName": "string",
                    "FatherName": "string",
                    "BirthDate": "string",
                    "Gender": "string",
                    "CivilStatus": "string",
                    "EmploymentDate": "string",
                    "MilitaryStatus": "string",
                    "PostponementDate": "string",
                    "Address": "string",
                    "createdAt": "string",
                    "updatedAt": "string",
                    "RoleId": 0,
                    "Role": {
                        "Id": 0,
                        "Name": "string",
                        "createdAt": "string",
                        "updatedAt": "string"
                    },
                    "CountryId": "string",
                    "Country": {
                        "Id": "string",
                        "Name": "string",
                        "createdAt": "string",
                        "updatedAt": "string"
                    },
                    "CityId": "string",
                    "City": {
                        "Id": "string",
                        "Name": "string",
                        "createdAt": "string",
                        "updatedAt": "string",
                        "CountryId": "string",
                        "Country": {
                            "Id": "string",
                            "Name": "string",
                            "createdAt": "string",
                            "updatedAt": "string"
                        }
                    },
                    "DistrictId": "string",
                    "District": {
                        "Id": "string",
                        "Name": "string",
                        "createdAt": "string",
                        "updatedAt": "string",
                        "CityId": "string",
                        "City": {
                            "Id": "string",
                            "Name": "string",
                            "createdAt": "string",
                            "updatedAt": "string",
                            "CountryId": "string",
                            "Country": {
                                "Id": "string",
                                "Name": "string",
                                "createdAt": "string",
                                "updatedAt": "string"
                            }
                        }
                    }
                }
            },
            "Role": {
                "type": "object",
                "required": ["Name", "createdAt", "updatedAt"],
                "properties": {
                    "Id": { "type": "integer", "description": "The auto-generated id of the role" },
                    "Name": { "type": "string", "description": "The role's name" },
                    "createdAt": { "type": "string", "format": "date-time", "description": "The role's created date" },
                    "updatedAt": { "type": "string", "format": "date-time", "description": "The role's updated date" }
                },
                "example": {
                    "Id": 0,
                    "Name": "string",
                    "createdAt": "string",
                    "updatedAt": "string"
                }
            },
            "Country": {
                "type": "object",
                "required": ["Id", "Name", "createdAt", "updatedAt"],
                "properties": {
                    "Id": { "type": "string", "description": "The id of the country" },
                    "Name": { "type": "string", "description": "The country's name" },
                    "createdAt": { "type": "string", "description": "The country's created date" },
                    "updatedAt": { "type": "string", "description": "The country's updated date" }
                },
                "example": {
                    "Id": "string",
                    "Name": "string",
                    "createdAt": "string",
                    "updatedAt": "string"
                }
            },
            "City": {
                "type": "object",
                "required": ["Id", "Name", "CountryId", "createdAt", "updatedAt"],
                "properties": {
                    "Id": { "type": "string", "description": "The id of the city" },
                    "Name": { "type": "string", "description": "The city's name" },
                    "createdAt": { "type": "string", "description": "The city's created date" },
                    "updatedAt": { "type": "string", "description": "The city's updated date" },
                    "CountryId": { "type": "string", "description": "The city's country id" },
                    "Country": { "$ref": "#/components/schemas/Country" }
                },
                "example": {
                    "Id": "string",
                    "Name": "string",
                    "createdAt": "string",
                    "updatedAt": "string",
                    "CountryId": "string",
                    "Country": {
                        "Id": "string",
                        "Name": "string",
                        "createdAt": "string",
                        "updatedAt": "string"
                    }
                }
            },
            "District": {
                "type": "object",
                "required": ["Id", "Name", "createdAt", "updatedAt", "CityId"],
                "properties": {
                    "Id": { "type": "string", "description": "The id of the district" },
                    "Name": { "type": "string", "description": "The district's name" },
                    "createdAt": { "type": "string", "description": "The district's created date" },
                    "updatedAt": { "type": "string", "description": "The district's updated date" },
                    "CityId": { "type": "string", "description": "The district's city id" },
                    "City": { "$ref": "#/components/schemas/City" }
                },
                "example": {
                    "Id": "string",
                    "Name": "string",
                    "createdAt": "string",
                    "updatedAt": "string",
                    "CityId": "string",
                    "City": {
                        "Id": "string",
                        "Name": "string",
                        "createdAt": "string",
                        "updatedAt": "string",
                        "CountryId": "string",
                        "Country": {
                            "Id": "string",
                            "Name": "string",
                            "createdAt": "string",
                            "updatedAt": "string"
                        }
                    }
                }
            },
            "CreateUpdateDocument": {
                "type": "object",
                "required": ["Name","FilePath","Status", "UserId"],
                "properties": {
                    "Id": { "type": "integer", "description": "The auto-generated id of the document" },
                    "Name": { "type": "string", "description": "The document's name" },
                    "FilePath": { "type": "string", "description": "The document's file path" },
                    "Status": { "type": "string", "description": "The document's status" },
                    "UserId": { "type": "integer", "description": "The document's user id" }
                }
            },
...
        }
    },
    "paths": {
...
        "/api/documents/GetAllDocumentsByUserId/{userId}": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "parameters": [{
                    "in": "path",
                    "name": "userId",
                    "description": "User Id",
                    "schema": {
                        "type": "integer"
                    },
                    "required": true
                }],
                "operationId": "allDocumentsByUserId",
                "responses": {
                    "200": {
                        "description": "Retrives a list of documents by user id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": { "$ref": "#/components/schemas/Document" }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/documents/GetDocument/{id}": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "parameters": [{
                    "in": "path",
                    "name": "id",
                    "description": "Document Id",
                    "schema": {
                        "type": "integer"
                    },
                    "required": true
                }],
                "operationId": "documentById",
                "responses": {
                    "200": {
                        "description": "Retrives a document by id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Document"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/documents/CreateDocument": {
            "post": {
                "tags": ["Documents"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUpdateDocument"
                            }
                        }
                    }
                },
                "operationId": "createDocument",
                "responses": {
                    "200": {
                        "description": "Create a new document data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Document"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/documents/EditDocument": {
            "put": {
                "tags": ["Documents"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUpdateDocument"
                            }
                        }
                    }
                },
                "operationId": "editDocument",
                "responses": {
                    "200": {
                        "description": "Edit an existing document data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Document"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/documents/DeleteDocument/{id}": {
            "delete": {
                "tags": ["Documents"],
                "parameters": [{
                    "in": "path",
                    "name": "id",
                    "description": "Document Id",
                    "schema": {
                        "type": "integer"
                    },
                    "required": true
                }],
                "operationId": "deleteDocument",
                "responses": {
                    "204": {
                        "description": "Delete an existing document data"
                    }
                }
            }
        },
...
    },
...
}